refactor: 패키지 분리 및 리네이밍
