+++ src/main/java/racingcar/Application.java
@@ -1,9 +1,11 @@
 package racingcar;
 
+import racingcar.controller.RacingCarController;
+
 public class Application {
 
     public static void main(String[] args) {
-        RacingGame racingGame = new RacingGame();
+        RacingCarController racingGame = new RacingCarController();
         racingGame.race();
     }
 }
