+++ src/main/java/racingcar/view/RacingCarView.java
@@ -1,9 +1,11 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.domain.Car;
 
 import java.util.Scanner;
 import java.util.List;
 
-public class RacingGameView {
+public class RacingCarView {
 
     private static final String START_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
     private static final String TRIAL_MESSAGE = "시도할 회수는 몇회인가요?";
